LSPS1: Move getOrder call to store + feedback changes
